Extract shell host command handling
